move map_*_field to Metadata module that can be included
