+++ lib/search/models/record.rb
@@ -1,5 +1,6 @@
 module Search::Models::Record
 end
 
+require "search/models/record/metadata"
 require "search/models/record/catalog"
 require "search/models/record/onlinejournals"
